Refactor app shell into section components and shared UI helpers
